+++ src/ENDFtk/section/2/151/UnresolvedEnergyDependentFissionWidths/JValue/src/generateList.hpp
@@ -1,12 +1,12 @@
 static std::vector< double >
 generateList( double spin,
               double d,
-              int amun,
+              double amun,
               double gn,
               double gg,
               std::vector< double >&& gf ) {
 
-  std::vector< double > list{ d, spin, static_cast<double>( amun ),
+  std::vector< double > list{ d, spin,  amun,
                               gn, gg, 0.0 };
   list.insert( list.end(), gf.begin(), gf.end() );
   return list;
